Refactor: Move interactions to interaction.js and visuals to draw.js
